# api/docs/swagger.yaml
definitions:
  payment.PaymentCreateReq:
    properties:
      amount:
        type: number
      payment_method:
        type: string
      payment_status:
        type: string
      reservation_id:
        type: string
    type: object
  payment.PaymentGetByIdResp:
    properties:
      amount:
        type: number
      id:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      reservation_id:
        type: string
    type: object
  reservation.CheckTimeReq:
    properties:
      restauran_id:
        type: string
      time:
        type: string
    type: object
  reservation.CheckTimeResp:
    properties:
      is_booked:
        type: boolean
    type: object
...
  reservation.GetAllReservationRes:
    properties:
      reservation:
        items:
          $ref: '#/definitions/reservation.ReservationRes'
        type: array
    type: object
...
  reservation.Reservation:
    properties:
      id:
        type: string
      reservation_time:
        type: string
      restaurant_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
...
  reservation.ReservationReq:
    properties:
      reservation_time:
        type: string
      restaurant_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  reservation.ReservationRes:
    properties:
      id:
        type: string
      reservation_time:
        type: string
      restaurant:
        $ref: '#/definitions/reservation.Restaurant'
      status:
        type: string
      user:
        $ref: '#/definitions/reservation.UserResp'
    type: object
  reservation.Restaurant:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  reservation.UserResp:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /reservation:
    post:
      consumes:
      - application/json
      description: Create a new reservation
      parameters:
      - description: Reservation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/reservation.ReservationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.Reservation'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create reservation
      tags:
      - reservation
  /reservation/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservation deleted
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Reservation not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete reservation
      tags:
      - reservation
    get:
      consumes:
      - application/json
      description: Get a reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.ReservationRes'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Reservation not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get reservation
      tags:
      - reservation
    put:
      consumes:
      - application/json
      description: Update an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated reservation data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/reservation.ReservationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.ReservationReq'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Reservation not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update reservation
      tags:
      - reservation
  /reservation/check/:
    post:
      consumes:
      - application/json
      description: Check reservations
      parameters:
      - description: Check data
        in: body
        name: chesk
        required: true
        schema:
          $ref: '#/definitions/reservation.CheckTimeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.CheckTimeResp'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Check reservations
      tags:
      - reservation
...
  /reservations:
    get:
      consumes:
      - application/json
      description: Get all reservations
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.GetAllReservationRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all reservations
      tags:
      - reservation
  /reservations/{id}/payment:
    post:
      consumes:
      - application/json
      description: Process payment for a reservation
      parameters:
      - description: Payment data
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/payment.PaymentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.PaymentGetByIdResp'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Process payment for a reservation
      tags:
      - reservation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"